requirements.txt
gate-api==6.0.1
python-dotenv==1.0.1
pandas==2.2.2
